Cleared more VSWE ridiculousness

+++ src/main/java/advancedsystemsmanager/flow/elements/Variable.java
@@ -37,7 +37,7 @@ public class Variable implements IContainerSelection<GuiManager>
     }
 
     @Override
-    public int getId()
+    public long getId()
     {
         return id;
     }
